implement layout_to_sizeclass() in terms of sizeclass_to_slotsize()
Add some constants

I'm annoyed that Rust won't automatically convert u8's to usize's for me, and won't, without verbiage, let me do things like convert u128's to usize's while raising an exception if it is lossy. But, oh well, since I often need to use things as indexes into arrays/slices, then I'm just going to use usize for all my numbers.

+++ benches/benches.rs
@@ -60,7 +60,7 @@ fn bench_layout_to_sizeclass_hugealign(b: &mut Bencher) {
 #[bench]
 fn bench_sizeclass_to_slotsize(b: &mut Bencher) {
     let mut r = rand::rng();
-    let reqscs: Vec<u8> = (0..NUM_ARGS).map(|_| r.random_range(0..35)).collect();
+    let reqscs: Vec<usize> = (0..NUM_ARGS).map(|_| r.random_range(0..35)).collect();
     let mut i = 0;
 
     b.iter(|| {
